Work in progress for TeleControl using the new spindle
Spindle size has increased for the lift pulley system. This makes the
pulley system much faster, but also makes the turning moment high enough
to cause the pulley to rotate backward under load. This has to be
countered.
Also the touch sensor has been incorporated to prevent the thread from
breaking.  There are still several bugs in the movement of the lift.